requirements.txt
flask
ipython
python-dotenv
requests
pytest
